refactor: move Button out of /action-folder. Delete /action-folder
