--- @udir-design/react/src/action/index.ts
@@ -1,1 +1,0 @@
-export * from './button/Button';
