+++ @udir-design/react/src/button/Button.stories.tsx
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from '@storybook/react';
 import { within, expect } from '@storybook/test';
 import { Button } from './Button';
-import { spaceForFocusStyleDecorator } from '../../../.storybook/utils/spaceForFocusStyleDecorator';
+import { spaceForFocusStyleDecorator } from '../../.storybook/utils/spaceForFocusStyleDecorator';
 
 const meta: Meta<typeof Button> = {
   component: Button,
