+++ @udir-design/react/src/button/Snapshots.dynamic.tsx
@@ -1,8 +1,8 @@
 import type { StoryObj } from '@storybook/react';
 import { Button } from './Button';
 import { _Button } from './Button.stories';
-import { createActionStateStories } from '../../../.storybook/utils/createActionStateStories';
-import { spaceForFocusStyleDecorator } from '../../../.storybook/utils/spaceForFocusStyleDecorator';
+import { createActionStateStories } from '../../.storybook/utils/createActionStateStories';
+import { spaceForFocusStyleDecorator } from '../../.storybook/utils/spaceForFocusStyleDecorator';
 import type { GeneratorMeta } from '@internal/dynamic-csf';
 
 type Story = StoryObj<typeof Button>;
