+++ @udir-design/react/src/index.ts
@@ -1,1 +1,1 @@
-export * from './action';
+export * from './button/Button';
